Ensure fullscreen windows don't lose focus when changing tags
It is rather confusing to switch tags to one that has a fullscreen
window only to discover after multiple keypresses that another
window actually has focus.

So, when switching to a tag with a fullscreen window, that window
will now be focused. If there are 2 fullscreen windows then the first
will be focused and the second will be kindly asked to stop being
fullscreen because having 2 is simply rude.